refactor(inventory): extract page-fetch config builders to internal module
